refactor: simplify runtime staging verification
